requirements.txt
numpy
pandas
python-dotenv==1.0.1
pip
beautifulsoup4==4.12.3
chromadb==0.5.0
black==24.4.2
isort==5.13.2
pytest==8.2.0
langgraph-cli==0.1.35
tiktoken==0.7.0
langchain-openai==0.1.12
pypdf==4.3.1
faiss-cpu==1.8.0.post1
langchain==0.2.12
langchain-community==0.2.11
langchainhub==0.1.20
langgraph>=0.2.18,<0.3.0
langgraph-checkpoint-sqlite==1.0.0
tavily-python==0.4.0
langchain-chroma==0.1.3